requirements.txt
requests==2.23.0
ics==0.7
google.cloud.logging==1.15.0